Move BrickFragment up into the app
Fixes https://github.com/wayfair/brickkit-android/issues/107
