+++ BrickKit/app/src/main/java/com/wayfair/brickkitdemo/InfiniteScrollBrickFragment.java
@@ -9,7 +9,6 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import com.wayfair.brickkit.brick.BaseBrick;
-import com.wayfair.brickkit.BrickFragment;
 import com.wayfair.brickkit.padding.BrickPaddingFactory;
 import com.wayfair.brickkit.OnReachedItemAtPosition;
 import com.wayfair.brickkit.size.FullPhoneFullHalfTabletBrickSize;
