+++ BrickKit/app/src/main/java/com/wayfair/brickkitdemo/SimpleBrickFragment.java
@@ -5,7 +5,6 @@ package com.wayfair.brickkitdemo;
 
 import android.os.Bundle;
 
-import com.wayfair.brickkit.BrickFragment;
 import com.wayfair.brickkit.brick.BaseBrick;
 import com.wayfair.brickkit.brick.ViewModelBrick;
 import com.wayfair.brickkit.padding.BrickPaddingFactory;
