+++ BrickKit/app/src/main/res/layout/text_brick.xml
@@ -13,7 +13,7 @@
         android:textColor="@android:color/black"
         android:textSize="31sp"
         android:background="@android:color/white"
-        android:padding="@dimen/default_brick_content_padding"
+        android:padding="@dimen/eight_dp"
         tools:text="Some text. Some text. Some text. Some text. Some text. Some text. Some text. Some text. " />
 
 </RelativeLayout>
