+++ BrickKit/app/src/main/res/layout/text_brick_vm.xml
@@ -20,7 +20,7 @@
             android:textColor="@android:color/black"
             android:textSize="31sp"
             android:background="@android:color/white"
-            android:padding="@dimen/default_brick_content_padding"
+            android:padding="@dimen/eight_dp"
             android:text="@{textViewModel.text}"
             tools:text="Some text. Some text. Some text. Some text. Some text. Some text. Some text. Some text. " />
