+++ BrickKit/app/src/main/res/layout/unused_brick.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="60dp"
     android:background="@android:color/darker_gray"
-    android:padding="@dimen/default_brick_content_padding"
+    android:padding="@dimen/eight_dp"
     tools:ignore="Overdraw">
 
 </RelativeLayout>
