+++ BrickKit/app/src/main/res/layout/vertical_fragment_brick.xml
