+++ BrickKit/bricks/src/androidTest/res/layout/text_brick_vm_placeholder.xml
@@ -1,8 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!-- Copyright © 2017 Wayfair. All rights reserved. -->
-<layout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+<layout xmlns:android="http://schemas.android.com/apk/res/android">
+
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content" >
@@ -14,7 +13,7 @@
             android:textColor="@android:color/black"
             android:textSize="31sp"
             android:background="@android:color/darker_gray"
-            android:padding="@dimen/default_brick_content_padding" />
+            android:padding="8dp" />
 
     </RelativeLayout>
