--- BrickKit/bricks/src/main/res/values/dimen.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Copyright © 2017 Wayfair. All rights reserved. -->
-<resources>
-    <dimen name="default_brick_content_padding">8dp</dimen>
-    <dimen name="no_dp">0dp</dimen>
-</resources>
